requirements.txt
Django==3.2.22
environs==5.2.1
dj-database-url==0.5.0
django-phonenumber-field[phonenumberslite]==7.2.0